Function to change the state of the cell
With correct type definition of the state and dedicated, if silly change state, it's a step for issue #14, also error detection for silliness. I suspect there
would be multiple additional checks. Possibly need to move error definition outside for simplicity of checking and removal of future additional checks of state.